+++ CHANGELOG.md
@@ -60,6 +60,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Chess activity move history scrollbar bug #2145
 - Speak activity settings not saving properly #2139
 - Story Activity: Images not displayed and inconsistent text size in .doc export #2155
+- Wrong export format message in Curriculum activity #2146
 
 ## [1.9.0] - 2025-03-25
 ### Added